// src/common/common-module/src/enums/lender.enum.ts
export enum LenderCredentials {
    RAZORPAY_EMANDATE = 'RazorPay Emandate',
    RAZORPAY_PENNY = 'RazorPay Penny',
    DECENTRO_PENNY = 'Decentro Penny' ,
    EXPERIAN_HARD_PULL = 'Experian Hard Pull', 
    PAYU_REPAYMENT = 'Payu Repayment',
  }

  export enum LenderName {
      RAMFINCORP = 1,
      KAMAKSHI = 2
  }

  export enum LenderStatus{
    ACTIVE=1,
    DEACTIVE=0
  }